Factor out pos/range LSP conversions into separate module
