+++ src/LanguageServer/Handlers.hs
@@ -5,16 +5,16 @@ module LanguageServer.Handlers( handlers
 import Control.Lens                      ( (^.) )
 import Control.Monad.IO.Class            ( liftIO )
 import Common.Error
-import qualified Common.Range            as R
 import Compiler.Compile                  ( compileModuleOrFile, Terminal (..) )
 import Compiler.Options                  ( Flags )
 import Compiler.Module                   ( loadedModule, modRangeMap )
 import Data.Maybe                        ( fromJust )
 import qualified Data.Text               as T
 import Language.LSP.Server
-import Lib.PPrint                        ( Pretty (..) )
 import qualified Language.LSP.Types      as J
 import qualified Language.LSP.Types.Lens as J
+import LanguageServer.Conversions
+import Lib.PPrint                        ( Pretty (..) )
 import Syntax.Colorize                   ( signature )
 import Syntax.RangeMap                   ( rangeMapFindAt, RangeInfo (..), NameInfo (..) )
 import Type.Pretty                       ( defaultEnv )
@@ -63,22 +63,6 @@ formatHoverContents rinfo = case rinfo of
 fromRight' :: Show a => Either a b -> b
 fromRight' (Right x) = x
 fromRight' (Left e)  = error $ "fromRight' failed: " ++ show e
-
-toLspPos :: R.Pos -> J.Position
-toLspPos p = J.Position (R.posLine p - 1) (R.posColumn p - 1) -- LSP positions are zero-based
-
-toLspRange :: R.Range -> J.Range
-toLspRange r = J.Range (J.Position l1 c1) (J.Position l2 $ c2 + 1) -- LSP range ends are exclusive
-  where
-    J.Position l1 c1 = toLspPos $ R.rangeStart r
-    J.Position l2 c2 = toLspPos $ R.rangeEnd r
-
-fromLspPos :: J.Uri -> J.Position -> R.Pos
-fromLspPos uri (J.Position l c) = R.makePos src (-1) (l + 1) (c + 1)
-  where
-    src = case J.uriToFilePath uri of
-      Just filePath -> R.Source filePath R.bstringEmpty -- TODO: Read file here?
-      Nothing       -> R.sourceNull
 
 -- TODO: Emit messages via LSP's logging mechanism
 terminal :: Terminal
